refactor: extract ui, accounts, ledger modules
